requirements.txt
alembic==1.8.1
amqp==5.1.1
anyio==3.6.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
fastapi==0.79.0
FastAPI-SQLAlchemy==0.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
numpy==1.23.1
packaging==21.3
pandas==1.4.3
prompt-toolkit==3.0.30
psycopg2==2.9.3
pydantic==1.9.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing-extensions==4.3.0
uvicorn==0.18.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
